+++ packages/frontend/src/ResizeHandle.module.css
