+++ packages/frontend/src/features/search/components/Search.tsx
@@ -1,7 +1,7 @@
 import { Box } from "@mantine/core";
 import clsx from "clsx";
 import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
-import styles from "../../layout/components/ResizeHandle.module.css";
+import styles from "../../../ResizeHandle.module.css";
 import { SearchContent } from "./SearchContent";
 import { SearchNavigation } from "./SearchNavigation";
